Revert "feat: Tag List on Unit page (feature-flagged) (#33645)"
This reverts commit 104d42c67a829a02c97baf6098c4026e401a0897.
