+++ cms/djangoapps/contentstore/utils.py
@@ -1,7 +1,6 @@
 """
 Common utility functions useful throughout the contentstore
 """
-from __future__ import annotations
 import configparser
 import logging
 from collections import defaultdict
@@ -443,7 +442,7 @@ def get_taxonomy_list_url():
     return taxonomy_list_url
 
 
-def get_taxonomy_tags_widget_url(course_locator=None) -> str | None:
+def get_taxonomy_tags_widget_url(course_locator) -> str:
     """
     Gets course authoring microfrontend URL for taxonomy tags drawer widget view.
 
@@ -452,9 +451,7 @@ def get_taxonomy_tags_widget_url(course_locator=None) -> str | None:
     taxonomy_tags_widget_url = None
     # Uses the same waffle flag as taxonomy list page
     if use_tagging_taxonomy_list_page():
-        mfe_base_url = settings.COURSE_AUTHORING_MICROFRONTEND_URL
-        if course_locator:
-            mfe_base_url = get_course_authoring_url(course_locator)
+        mfe_base_url = get_course_authoring_url(course_locator)
         if mfe_base_url:
             taxonomy_tags_widget_url = f'{mfe_base_url}/tagging/components/widget/'
     return taxonomy_tags_widget_url
