+++ cms/static/js/models/xblock_info.js
@@ -173,10 +173,6 @@ define(
              * True if summary configuration is enabled.
              */
                 summary_configuration_enabled: null,
-               /**
-                * List of tags of the unit. This list is managed by the content_tagging module.
-                */
-               tags: null,
             },
 
             initialize: function() {
