+++ cms/templates/container.html
@@ -213,10 +213,7 @@ from openedx.core.djangolib.markup import HTML, Text
                                 )}</span>
                             </p>
                         </div>
-                    </div>
-                    % if show_unit_tags:
-                        <div class="unit-tags is-hidden"></div>
-                    % endif
+                     </div>
                 % endif
             </aside>
         </section>
@@ -241,7 +238,4 @@ from openedx.core.djangolib.markup import HTML, Text
         </div>
     </div>
 </div>
-
-<div id="manage-tags-drawer" class="drawer"></div>
-<div class="drawer-cover"></div>
 </%block>
